Update per Module 2.2.3
Updated the VBA code as instructed throughout the lesson. The latest revision asks to run a ticker on the DQ stock symbol. Because the lesson jumps from different code to code during the lesson, I have create sub routines for every spot that it deviates. The code it is asking to be turned in and ensure is working is within the sub routine titled "DQAnalysis5"